+++ frontend/src/__tests__/cypress/cypress/fixtures/e2e/dataScienceProjects/testWorkbenchEditing.yaml
@@ -3,3 +3,4 @@ editTestNamespace: 'dsp-wb-edit-test'
 editedTestNamespace: 'dsp-wb-edited-test'
 editedTestDescription: 'This is a test description.'
 pvc_EditDisplayName: 'pvc--edit-test-display-name'
+pvcStorageName: 'wb-edit-test-storage'
